BugFix: Remove redundant print message in Python script
Remove a redundant print message from the python script that was
apparently overlooked while making the previous commit.

Normally this wouldn't be much of a problem, but since the output
of the python script will be handled by bash, even a single redundant
print message can been a major source of problem.