Update: Add shell script to batch-add all team drives to RClone.
Add shell script that collects the names of the team drives, and their
ID's and then batch adds all these TeamDrives to RClone config file.

This script internally calls the Python script that will fetch a list
of all the team drives that are avaiable with the associated account.

The python script dumps the name of the team drive, and its ID in a text
file, the shell script then scrapes this text file to get the data.