Update: Create script to handle batch-mounting of configs.
Return negative value while breaking out of "add-configs.sh" to
indicate that the execution was terminated mid-way.

Add new script that will handle the process of batch mounting
all RClone configs. Asks the user to enter a path for the root
directory, or extracts the path of the root directory if
supplied as an argument and validates this path.