Update: Extract list of config(s), create directories for configs
Try to fetch the RClone configs file, if the file can't be located ask
the user to manually enter the location of the config file.

Parse the text from the config file, for each line that contains the name
of the config, create a new directory inside the root directory for the
particular config. If a directory with the same name already exists, print an
error message to the user.

Parse the text from the config file, for each line that contains the name
of the config, create a new directory inside the root directory for the
particular config. If a directory with the same name already exists, print an
error message to the user.

While creating directories for the configs, replace the forbidden characters
from the name of the directory with a space.

Convert indentation from spaces to tabs, allow the script to run without
root privileges.

(╯°□°）╯︵ ┻━┻