BugFix: Patch bug to prevent adding configs with invalid names.
Apparently, the only characters allowed in the names of a config are
alphabets (lower case and upper case), underscore, hyphen and spaces.

Anything outside of this set is an invalid character for a config name.

The python script never made any checks for this. Meaning that
team-drives having characters not allowed in config name will also be
added to RClone as a config.

This updates patches this bug, using regex reverse matching in Python,
any character that is not present in the set of allowed characters
will be replaced by a space.