Feature: Complete a basic working version of the script
Add code to create a new text file (or wipe out the existing file),
this text file will contain the full paths of the directories
where RClone configs are mounted by this script.

The data from this file will be used to create a script to unmount
all the mounted configs.

Add code to create a new directory inside the root directory for every
config and mount the config to the corresponding directory.

Minor improvements in the comments to better explain the working of the
script. Also, create a string containing a few hard-coded arguments that
will be used while mounting the config(s).