BugFix: Patch major bug that prevented mounting configs
This version is final working version of the main mount script.

The `if` condition in the previous commit was inverted, this meant
that while reading the config file, the actual name of the config
will be skipped and any other line would be treated as the name of
the config. Due to this, none of the configs could be mounted.

I should've tested the change before making a commit -- and marking
the commit as a feature (O_O;)