Feature: Allow force mounting, additional check(s)
Before mounting the directory, perform a check to ensure that the
directory to be mounted actually exists. Upto now, the script was
checking if the root directory exists (this was the bug,
the check for root directory was to be performed separately).

Also, this update finally uses the force mount if the argument is
supplied. This means that if a directory with the same name as the
config already exists inside the root directory, the existing
directory will be forcefully mounted even if it is not empty.

Improve comments to make the code more readable (and less complex).