Update: Add check to ensure a config is mounted.
Add a check to ensure that the config is mounted, allow the user to
make the script sleep for a fixed amount of time before checking if
the background process is still running.

If the process is not running by the time the sleep ends, it means
that the config could not be mounted due to an error and the process
is now dead.