Feature: Add script to unmount all the mounted configs at once.
((Because unmounting the configs one-by-one is too mainstream :p))

Add a script that will unmount all the configs mounted by
'multi-mount.sh'

This script uses the same text file to get the paths to the
directories mounted, and then uses fusermount to unmount all the
directories. Once this is done, the mount file is wiped clean.

This scipt can also optionally delete the mounted directories instead
of simply unmounting them if required, and can also suppress any error
messages with the correct arguments.