Feature: Allow renaming config before mounting it
Allow customizing the titles of the directories to which each config
is mounted.

Add sample configuration file containing information on how to add data
to the file so that the script can parse the data and mount accordingly.

Minor changes to the sample configuration file for ignore list, attempt
explaining stuff in a clear(er) manner.

Minor changes made to code documentation to tidy up the code and clearly
explain the logic.