Minor: Add requirements. Change the layout of the root repo.
Add requirements required for `list_generator.py` to fetch a list of
team-drives that an account has access to.

Move the files around to ensure that the root of the repo contains only
those files that are directly related to the repository.

Update `.gitignore` to not ignore all text files by default.